# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * kode.live API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface Project {
  id: string;
  name: string;
  /** @nullable */
  description: string | null;
  status: ProjectStatus;
  messageCount?: number;
  taskCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  status?: ProjectUpdateStatus;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: string;
  projectId: string;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export type MessageInputRole = typeof MessageInputRole[keyof typeof MessageInputRole];


export const MessageInputRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface MessageInput {
  role: MessageInputRole;
  content: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface Task {
  id: string;
  projectId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  order: number;
  createdAt: string;
}

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  order?: number;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  order?: number;
}

export interface DashboardStats {
  totalProjects: number;
  activeProjects: number;
  completedProjects: number;
  totalMessages: number;
  totalTasks: number;
}