# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: kode.live API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project operations
  - name: messages
    description: Message/conversation operations
  - name: tasks
    description: Planner task operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /projects/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Send a message in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /projects/{id}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List planner tasks in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a planner task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /projects/{id}/tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a planner task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a planner task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /stats:
    get:
      operationId: getStats
      tags: [projects]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Project:
      type: object
      required: [id, name, description, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, archived]
        messageCount:
          type: number
        taskCount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [active, completed, archived]

    Message:
      type: object
      required: [id, projectId, role, content, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    MessageInput:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string

    Task:
      type: object
      required: [id, projectId, title, status, order, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, done]
        order:
          type: number
        createdAt:
          type: string
          format: date-time

    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        order:
          type: number

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, done]
        order:
          type: number

    DashboardStats:
      type: object
      required: [totalProjects, activeProjects, completedProjects, totalMessages, totalTasks]
      properties:
        totalProjects:
          type: number
        activeProjects:
          type: number
        completedProjects:
          type: number
        totalMessages:
          type: number
        totalTasks:
          type: number