# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * kode.live API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalProjects: number;
  activeProjects: number;
  completedProjects: number;
  totalMessages: number;
  totalTasks: number;
}