# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * kode.live API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageInputRole } from './messageInputRole';

export interface MessageInput {
  role: MessageInputRole;
  content: string;
}