# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * kode.live API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectUpdateStatus } from './projectUpdateStatus';

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  status?: ProjectUpdateStatus;
}