# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * kode.live API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: string;
  projectId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  order: number;
  createdAt: Date;
}