# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * kode.live API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  order?: number;
}